+++ .serverless/Template.myNextApplication.CloudFront.json
@@ -26,7 +26,7 @@
             "PATCH"
           ],
           "lambda@edge": {
-            "origin-request": "arn:aws:lambda:us-east-1:114186083917:function:ej3l148-x0ewho8:118"
+            "origin-request": "arn:aws:lambda:us-east-1:114186083917:function:ej3l148-x0ewho8:120"
           }
         }
       }
@@ -40,7 +40,7 @@
       "queryString": true
     },
     "lambda@edge": {
-      "origin-request": "arn:aws:lambda:us-east-1:114186083917:function:ej3l148-rt3xvs9:326"
+      "origin-request": "arn:aws:lambda:us-east-1:114186083917:function:ej3l148-rt3xvs9:328"
     }
   }
 }
