+++ .serverless/Template.myNextApplication.apiEdgeLambda.json
@@ -1,6 +1,6 @@
 {
   "name": "ej3l148-x0ewho8",
-  "hash": "rZ/ttYwbi7cU2+b+cNDknsaPKcX3uA0I6peKvPlz+Tk=",
+  "hash": "axgxcUB0W7W2xltDNlYTgy7HKv6Sd72cTfRm3TzoYAs=",
   "description": "undefined (API)",
   "memory": 512,
   "timeout": 10,
